be able to build stuff

+++ IntegrationTests/IntegrationTests/Steps/Records.cs
@@ -1,7 +1,7 @@
 ﻿namespace IntegrationTests.Steps
 {
     public record Block(Coordinate Coordinate, string Name, string BlockType);
-    public record Coordinate(int X, int Y, int Z);
+    public record Coordinate(float X, float Y, float Z);
 
-    public record Vector(int X, int Y, int Z);
+    public record Vector(float X, float Y, float Z);
 }
